Add storybook-addon-designs which allows embedding links to Figma in stories
