+++ packages/synapse-react-client/.storybook/main.ts
@@ -10,6 +10,7 @@ const config: StorybookConfig = {
     '@storybook/addon-links',
     '@storybook/addon-essentials',
     '@storybook/addon-interactions',
+    'storybook-addon-designs',
   ],
   framework: {
     name: '@storybook/react-vite',
