+++ packages/synapse-react-client/package.json
@@ -44,8 +44,8 @@
     "@react-hook/resize-observer": "^1.2.6",
     "@rjsf/core": "^5.7.3",
     "@rjsf/mui": "^5.7.3",
-    "@rjsf/validator-ajv8": "^5.7.3",
     "@rjsf/utils": "^5.7.3",
+    "@rjsf/validator-ajv8": "^5.7.3",
     "@sage-bionetworks/react-base-table": "^1.13.2",
     "@types/tinycolor2": "^1.4.3",
     "@types/ua-parser-js": "^0.7.36",
@@ -219,6 +219,7 @@
     "react-test-renderer": "^18.2.0",
     "rollup-plugin-polyfill-node": "^0.10.2",
     "storybook": "^7.0.17",
+    "storybook-addon-designs": "7.0.0-beta.2",
     "stream-browserify": "^3.0.0",
     "stream-http": "^3.2.0",
     "timers-browserify": "^2.0.12",
