--- packages/synapse-react-client/stories/stubs/Button/MuiButton.stories.ts
@@ -1,17 +1,0 @@
-import { Meta, StoryObj } from '@storybook/react'
-import { Button } from './MuiButton'
-
-const meta = {
-  title: 'UI/MUI/Button',
-  component: Button,
-} satisfies Meta
-export default meta
-
-type Story = StoryObj<typeof meta>
-
-export const Demo: Story = {
-  args: {
-    variant: 'contained',
-    label: 'My Button Text Here',
-  },
-}
