+++ packages/synapse-react-client/stories/stubs/TextField/MuiTextField.stories.ts
@@ -4,6 +4,12 @@ import { TextField } from './MuiTextField'
 const meta = {
   title: 'UI/MUI/TextField',
   component: TextField,
+  parameters: {
+    design: {
+      type: 'figma',
+      url: 'https://www.figma.com/file/0oPm5lLSUva8kyfVNMS6FA/Sage-Style-%26-Component-Library?type=design&node-id=540-19147',
+    },
+  },
 } satisfies Meta
 export default meta
